PKG -- [fcl] shared exports
